(bug) fix foreign key relations and generate seed data

+++ backend/models/common.go
@@ -5,7 +5,7 @@ import (
 )
 
 type CommonModel struct {
-	ID        uint       `gorm:"primary_key" json:"id"`
+	ID        uint8      `gorm:"primary_key" json:"id"`
 	CreatedAt time.Time  `json:"created_at"`
 	UpdatedAt time.Time  `json:"updated_at"`
 	DeletedAt *time.Time `json:"deleted_at"`
